+++ Editor/Attributes/HideInPlayModeDrawer.cs
@@ -1,0 +1,37 @@
+using UnityEditor;
+using UnityEngine;
+
+namespace AnkleBreaker.Utils.Inspector.Editor
+{
+    [CustomPropertyDrawer(typeof(HideInPlayModeAttribute))]
+    public class HideInPlayModeDrawer : PropertyDrawer
+    {
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            if (Application.isPlaying) return;
+            EditorGUI.PropertyField(position, property, label, true);
+        }
+
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            if (Application.isPlaying) return 0f;
+            return EditorGUI.GetPropertyHeight(property, label, true);
+        }
+    }
+
+    [CustomPropertyDrawer(typeof(HideInEditorModeAttribute))]
+    public class HideInEditorModeDrawer : PropertyDrawer
+    {
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            if (!Application.isPlaying) return;
+            EditorGUI.PropertyField(position, property, label, true);
+        }
+
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+        {
+            if (!Application.isPlaying) return 0f;
+            return EditorGUI.GetPropertyHeight(property, label, true);
+        }
+    }
+}